[Fix] fix compilation error under bsd

+++ ladder/include/EventPoller.h
@@ -55,8 +55,8 @@ private:
   int cur_poll_size_;
   PipePtr pipe_;
 #ifdef __FreeBSD__
-	static std::map<short, int64> flt_2_stat_;	// filter to status
-	static std::map<int64, short> stat_2_flt_;			// status to filter
+	static std::map<short, uint32_t> flt_2_stat_;	// filter to status
+	static std::map<uint32_t, short> stat_2_flt_;			// status to filter
 #endif
 };
